// box.css.d.ts
export declare const contentSizeVariants: {
    height: {
        xxsmall: {
            height: `var(--${string})` | `var(--${string}, ${string})` | `var(--${string}, ${number})`;
        };
        xsmall: {
            height: `var(--${string})` | `var(--${string}, ${string})` | `var(--${string}, ${number})`;
        };
        small: {
            height: `var(--${string})` | `var(--${string}, ${string})` | `var(--${string}, ${number})`;
        };
        medium: {
            height: `var(--${string})` | `var(--${string}, ${string})` | `var(--${string}, ${number})`;
        };
        large: {
            height: `var(--${string})` | `var(--${string}, ${string})` | `var(--${string}, ${number})`;
        };
        xlarge: {
            height: `var(--${string})` | `var(--${string}, ${string})` | `var(--${string}, ${number})`;
        };
        none: {
            height: string;
        };
    };
    width: {
        xxsmall: {
            width: `var(--${string})` | `var(--${string}, ${string})` | `var(--${string}, ${number})`;
        };
        xsmall: {
            width: `var(--${string})` | `var(--${string}, ${string})` | `var(--${string}, ${number})`;
        };
        small: {
            width: `var(--${string})` | `var(--${string}, ${string})` | `var(--${string}, ${number})`;
        };
        medium: {
            width: `var(--${string})` | `var(--${string}, ${string})` | `var(--${string}, ${number})`;
        };
        large: {
            width: `var(--${string})` | `var(--${string}, ${string})` | `var(--${string}, ${number})`;
        };
        xlarge: {
            width: `var(--${string})` | `var(--${string}, ${string})` | `var(--${string}, ${number})`;
        };
        none: {
            width: string;
        };
    };
};
export declare const gapVariants: {
    gap: {
        xxsmall: {
            gap: `var(--${string})` | `var(--${string}, ${string})` | `var(--${string}, ${number})`;
        };
        xsmall: {
            gap: `var(--${string})` | `var(--${string}, ${string})` | `var(--${string}, ${number})`;
        };
        small: {
            gap: `var(--${string})` | `var(--${string}, ${string})` | `var(--${string}, ${number})`;
        };
        medium: {
            gap: `var(--${string})` | `var(--${string}, ${string})` | `var(--${string}, ${number})`;
        };
        large: {
            gap: `var(--${string})` | `var(--${string}, ${string})` | `var(--${string}, ${number})`;
        };
        xlarge: {
            gap: `var(--${string})` | `var(--${string}, ${string})` | `var(--${string}, ${number})`;
        };
        none: {
            gap: number;
        };
    };
};